requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
transformers>=4.31.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.0.0
